Change timeline to fit funding schedule
The DIANA/HEP fellowship funds are controlled by the University of
Cincinnati and are treated as per diem. As Matthew was travelling for
the 2nd S2I2 HEP/CS Workshop for the first week of May and also
recieving per diem from the University of Cincinnati through the S2I2
grant the start and end dates of the fellowship were moved back one week to avoid
overlap.